requirements.txt
Django>=2.1.3,<3.8.0
djangorestframework>=3.9.0,<4.10.0
djangorestframework-simplejwt==4.4.0
django-rest-auth==0.9.5
rest_framework_jwt
requests==2.24.0